feat(#6 綁定 redux store):
