+++ packages/splitbill-web/src/screens/HomeScreen/index.js
@@ -1,3 +1,8 @@
-export default function HomeScreen() {
-  return <h3>Home Screen</h3>;
-}
+import { connect } from 'react-redux';
+import HomeScreen from './view';
+
+const mapStateToProps = ({ auth }) => ({ auth });
+
+const mapDispatchToProps = () => ({});
+
+export default connect(mapStateToProps, mapDispatchToProps)(HomeScreen);
