+++ packages/splitbill-web/src/screens/HomeScreen/view.js
@@ -1,0 +1,8 @@
+const HomeScreen = (props) => {
+  return (<div>
+    <h3>Home Screen</h3>
+    <p>auth: {JSON.stringify(props.auth)}</p>
+  </div>);
+}
+
+export default HomeScreen;
