+++ packages/splitbill-web/src/screens/LoginScreen/index.js
@@ -1,43 +1,8 @@
-import {
-  useNavigate,
-  useLocation,
-} from "react-router-dom";
-import { useAuth } from "../../navigations";
+import { connect } from "react-redux";
+import LoginScreen from "./view";
 
-export default function LoginScreen() {
-  let navigate = useNavigate();
-  let location = useLocation();
-  let auth = useAuth();
+const mapStateToProps = () => ({aaa: "aaa"});
 
-  let from = location.state?.from?.pathname || "/";
+const mapDispatchToProps = () => ({aaafunc: () => false});
 
-  function handleSubmit(event) {
-    event.preventDefault();
-
-    let formData = new FormData(event.currentTarget);
-    let username = formData.get("username");
-
-    auth.signin(username, () => {
-      // Send them back to the page they tried to visit when they were
-      // redirected to the login page. Use { replace: true } so we don't create
-      // another entry in the history stack for the login page.  This means that
-      // when they get to the protected page and click the back button, they
-      // won't end up back on the login page, which is also really nice for the
-      // user experience.
-      navigate(from, { replace: true });
-    });
-  }
-
-  return (
-    <div>
-      <p>You must log in to view the page at {from}</p>
-
-      <form onSubmit={handleSubmit}>
-        <label>
-          Username: <input name="username" type="text" />
-        </label>{" "}
-        <button type="submit">Login</button>
-      </form>
-    </div>
-  );
-}
+export default connect(mapStateToProps, mapDispatchToProps)(LoginScreen);
