+++ packages/splitbill-web/src/screens/PublicScreen/index.js
@@ -1,3 +1,10 @@
-export function PublicScreen() {
-  return <h3>Public Screen</h3>;
-}
+import { connect } from 'react-redux';
+import PublicScreen from './view';
+
+const mapStateToProps = ({ auth }) => ({
+  auth,
+});
+
+const mapDispatchToProps = () => ({});
+
+export default connect(mapStateToProps, mapDispatchToProps)(PublicScreen);
